+++ plugins/healthcheck/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @internal/plugin-healthcheck
 
+## 0.1.2
+
+### Patch Changes
+
+- ad0fde9: Upgrade backstage to 1.7 and bump effection dependencies
+
 ## 0.1.1
 
 ### Patch Changes