+++ plugins/humanitec-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @frontside/backstage-plugin-humanitec-backend
 
+## 0.3.2
+
+### Patch Changes
+
+- ad0fde9: Upgrade backstage to 1.7 and bump effection dependencies
+
 ## 0.3.1
 
 ### Patch Changes